refactor: Migrated resource.ts
- Removed namespaces
- Added imports
- @TODO: Documentation
- @TODO: Check if code needs to be migrated
